requirements.txt
# Core dependencies for market data collection
requests>=2.31.0
pyyaml>=6.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
lxml>=4.9.0
